Avoid initializing free disk space in redis for all hue shell commands
